Various commits to bring it up-to-date
Vagrantfile was updated to accomodate for DNS usage when VPN interferes with name resolution. You have to vb.customize ["modifyvm", :id, "--natdnshostresolver1", "on"]

ansible.yml was updated to modify /etc/hosts to give a name to 192.168.98.100. This is because otherwise the last job for RKE failed since it could not resolv the IP address back to the name rkube

cluster.yml was updated to allow for a cluster name to be set